requirements.txt
PyAudio==0.2.14
torch
torchvision
torchaudio
openai-whisper
pydub==0.25.1
